Refactor: split UserServlet into separate servlets for CRUD operations
