+++ src/main/java/com/codegym/web/servlets/UserListServlet.java
@@ -1,0 +1,24 @@
+package com.codegym.web.servlets;
+
+import com.codegym.web.repository.InMemoryUserRepository;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+import java.io.IOException;
+
+public class UserListServlet extends HttpServlet {
+
+    private final InMemoryUserRepository userRepository;
+
+    public UserListServlet(InMemoryUserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        req.setAttribute("users", userRepository.findAll());
+        req.getRequestDispatcher("/WEB-INF/jsp/users.jsp").forward(req, resp);
+    }
+}
